Renamed all of the page files to a_b case.
