+++ apk_iman_ba/lib/pages/details_page.dart
